+++ services/cd-service/pkg/repository/repository.go
@@ -870,7 +870,7 @@ func (s *State) GetEnvironmentApplicationVersion(ctx context.Context, transactio
 	if depl == nil || depl.ReleaseNumbers.Version == nil {
 		return nil, nil
 	}
-	var v = uint64(*depl.ReleaseNumbers.Version)
+	var v = *depl.ReleaseNumbers.Version
 	return &v, nil
 }
 
@@ -1174,7 +1174,7 @@ func (s *State) GetApplicationRelease(ctx context.Context, transaction *sql.Tx, 
 		return nil, nil
 	}
 	return &Release{
-		Version:         uint64(*env.ReleaseNumbers.Version),
+		Version:         *env.ReleaseNumbers.Version,
 		UndeployVersion: env.Metadata.UndeployVersion,
 		SourceAuthor:    env.Metadata.SourceAuthor,
 		SourceCommitId:  env.Metadata.SourceCommitId,
